Replace env.json import with env.ts for type coverage
