+++ src/components/scenes/LoginScene.tsx
@@ -7,11 +7,11 @@ import { ImageSourcePropType, Keyboard, StatusBar, View } from 'react-native'
 import { checkVersion } from 'react-native-check-version'
 import { BlurView } from 'rn-id-blurview'
 
-import ENV from '../../../env.json'
 import { showSendLogsModal } from '../../actions/LogActions'
 import { initializeAccount, logoutRequest } from '../../actions/LoginActions'
 import { serverSettingsToNotificationSettings, setDeviceSettings } from '../../actions/NotificationActions'
 import { cacheStyles, Theme, ThemeProps, withTheme } from '../../components/services/ThemeContext'
+import { ENV } from '../../env'
 import s from '../../locales/strings'
 import { config } from '../../theme/appConfig'
 import { DeepLink } from '../../types/DeepLinkTypes'
