+++ src/components/services/Providers.tsx
@@ -7,8 +7,8 @@ import { Provider } from 'react-redux'
 import { applyMiddleware, createStore } from 'redux'
 import thunk from 'redux-thunk'
 
-import ENV from '../../../env.json'
 import { loadDeviceReferral } from '../../actions/DeviceReferralActions'
+import { ENV } from '../../env'
 import { rootReducer } from '../../reducers/RootReducer'
 import { Dispatch, RootState, Store } from '../../types/reduxTypes'
 import { errorAlert } from '../../util/middleware/errorAlert'
